Fixes #75 by using ignition ecm instead of traversing sdformat parents (#80)
* Fixes #75 by using ignition ecm instead of traversing sdformat parents

This PR addresses issue #75. @luca-della-vedova has found an alternative two liner fix in upstream sdformat that would solve our problems. However, to future proof the ignition version of the plugin we should try to use the ecm to retrieve properties instead of relying on sdformat for traversing the joint properties.

Signed-off-by: Arjo Chakravarty <arjo@openrobotics.org>
Co-authored-by: Luca Della Vedova <luca@openrobotics.org>