Qt5 and Qt6 compatibility (#157)
## Bug fix

### Fixed bug

Closes #155 

### Fix applied

Follows the approach proposed in https://github.com/open-rmf/rmf_simulation/issues/155#issuecomment-3376342767, we can:

* Remove Qt dependencies and just use the Qt version that is brought in by `gz-gui`.
* Use unversioned functions for CMake.
* Check at the `qml` level for the Qt version and adapt the class names with a leading underscore for qt >= 6.
* Remove `LightTuning` altogether since most of its content is now available in upstream Gazebo, it has also been untested for a long time since it's not part of the default plugins but I can try to readd it if there is need for it.

I tested this in a 24.04 distrobox with either rolling or jazzy and it seems to work OK, but happy to get another pair of eyes.
Note that rolling has a few other issues, https://github.com/open-rmf/rmf_visualization/issues/87 and gazebo failing to run (at least for me) because of an invalid command. I had to run Gazebo manually instead of using the launch file:

```
$: export GZ_SIM_RESOURCE_PATH=~/rmf_ws/install/rmf_demos_assets/share/rmf_demos_assets/models:~/rmf_ws/install/rmf_demos_maps/share/rmf_demos_maps/maps/hotel/models/
$: gz sim -r -v 3 /usr/local/google/home/lucadv/noble_ws/install/rmf_demos_maps/share/rmf_demos_maps/maps/hotel/hotel.world -z 100
```
---------

Signed-off-by: Luca Della Vedova <lucadv@intrinsic.ai>